+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 0.3.0 (unreleased)
+
+- Removed unnecessary requirements
+- Removed "copy to clipboard" feature that only works on OSX
+- Upgrade to Click 7
+- Fixed argument case inconsistency between pip and pipup
+
 ## Version 0.2.0 released June 25th, 2016
 
 - Fix unhelpful error message when pipup can't find your requirements.txt file. Thanks Melissa Hill for finding and reporting the bug